requirements.txt
# Geospatial processing
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0 
pyproj>=3.6.0

# Deep Learning PyTorch (CPU optimized)
torch>=2.0.0 
torchvision>=0.15.0 
segmentation-models-pytorch>=0.3.3

# Image processing
opencv-python-headless>=4.8.0 
albumentations>=1.3.0 
Pillow>=10.0.0 
scikit-image>=0.21.0

# Data handling
numpy>=1.24.0 
pandas>=2.0.0 
tqdm>=4.65.0

# Model optimization 
onnx>=1.14.0 
onnxruntime>=1.15.0

# Utilities 
pyyaml>=6.0 
matplotlib>=3.7.0 
seaborn>=0.12.0

# Memory profiling 
psutil>=5.9.0 
memory-profiler>=0.61.0

# Optional: For annotation (install separately if needed)
# label-studio>=1.9.0
scikit-learn>=1.8.0